use a single controller for provisioning and cleanup
