--- internal/controllers/s3user/controller.go
@@ -1,33 +1,0 @@
-/*
-Copyright 2023.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package s3user
-
-import (
-	ctrl "sigs.k8s.io/controller-runtime"
-	"sigs.k8s.io/controller-runtime/pkg/builder"
-
-	s3v1alpha1 "github.com/snapp-incubator/s3-operator/api/v1alpha1"
-	"github.com/snapp-incubator/s3-operator/internal/predicates"
-)
-
-func (r *Reconciler) SetupWithManager(mgr ctrl.Manager) error {
-	s3UserClassPredicate := predicates.NewS3ClassPredicate(r.s3UserClass)
-
-	return ctrl.NewControllerManagedBy(mgr).
-		For(&s3v1alpha1.S3User{}, builder.WithPredicates(s3UserClassPredicate)).
-		Complete(r)
-}
